fix eval hooks
