+++ mmdet/core/utils/__init__.py
@@ -1,12 +1,9 @@
 from .dist_utils import (init_dist, reduce_grads, DistOptimizerHook,
                          DistSamplerSeedHook)
-from .hooks import (EmptyCacheHook, DistEvalHook, DistEvalRecallHook,
-                    CocoDistEvalmAPHook)
-from .misc import tensor2imgs, unmap, results2json, multi_apply
+from .hooks import EmptyCacheHook
+from .misc import tensor2imgs, unmap, multi_apply
 
 __all__ = [
     'init_dist', 'reduce_grads', 'DistOptimizerHook', 'DistSamplerSeedHook',
-    'EmptyCacheHook', 'DistEvalHook', 'DistEvalRecallHook',
-    'CocoDistEvalmAPHook', 'tensor2imgs', 'unmap', 'results2json',
-    'multi_apply'
+    'EmptyCacheHook', 'tensor2imgs', 'unmap', 'multi_apply'
 ]
