+++ tools/configs/r50_fpn_maskrcnn_1x.py
@@ -106,16 +106,26 @@ data = dict(
         flip_ratio=0.5,
         with_mask=True,
         with_crowd=True,
-        with_label=True,
-        test_mode=False),
+        with_label=True),
+    val=dict(
+        type=dataset_type,
+        ann_file=data_root + 'annotations/instances_val2017.json',
+        img_prefix=data_root + 'val2017/',
+        img_scale=(1333, 800),
+        img_norm_cfg=img_norm_cfg,
+        size_divisor=32,
+        flip_ratio=0,
+        with_mask=True,
+        with_crowd=True,
+        with_label=True),
     test=dict(
         type=dataset_type,
         ann_file=data_root + 'annotations/instances_val2017.json',
         img_prefix=data_root + 'val2017/',
         img_scale=(1333, 800),
-        flip_ratio=0,
         img_norm_cfg=img_norm_cfg,
         size_divisor=32,
+        flip_ratio=0,
         with_mask=False,
         with_label=False,
         test_mode=True))
@@ -141,7 +151,7 @@ log_config = dict(
 # runtime settings
 total_epochs = 12
 device_ids = range(8)
-dist_params = dict(backend='nccl', port='29500')
+dist_params = dict(backend='gloo')
 log_level = 'INFO'
 work_dir = './work_dirs/fpn_mask_rcnn_r50_1x'
 load_from = None
