+++ tools/configs/r50_fpn_rpn_1x.py
@@ -65,16 +65,26 @@ data = dict(
         flip_ratio=0.5,
         with_mask=False,
         with_crowd=False,
-        with_label=False,
-        test_mode=False),
+        with_label=False),
+    val=dict(
+        type=dataset_type,
+        ann_file=data_root + 'annotations/instances_val2017.json',
+        img_prefix=data_root + 'val2017/',
+        img_scale=(1333, 800),
+        img_norm_cfg=img_norm_cfg,
+        size_divisor=32,
+        flip_ratio=0,
+        with_mask=False,
+        with_crowd=False,
+        with_label=False),
     test=dict(
         type=dataset_type,
         ann_file=data_root + 'annotations/instances_val2017.json',
         img_prefix=data_root + 'val2017/',
         img_scale=(1333, 800),
-        flip_ratio=0,
         img_norm_cfg=img_norm_cfg,
         size_divisor=32,
+        flip_ratio=0,
         with_mask=False,
         with_label=False,
         test_mode=True))
@@ -103,5 +113,5 @@ dist_params = dict(backend='gloo')
 log_level = 'INFO'
 work_dir = './work_dirs/fpn_rpn_r50_1x'
 load_from = None
-resume_from = None 
-workflow = [('train', 1)]
+resume_from = None
+workflow = [('train', 1), ('val', 1)]
